judge_preference/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.9.0
attrs==25.3.0
autobahn==24.4.2
automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.5
datasets==4.0.0
dill==0.3.8
distro==1.9.0
faiss-cpu==1.8.0
filelock==3.18.0
Flask==3.0.3
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.1
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iterable-io==1.0.0
jiter==0.10.0
magic-wormhole==0.19.2
matplotlib==3.9.2
multidict==6.6.3
multiprocess==0.70.16
numpy==2.2.2
omegaconf==2.3.0
openai==1.97.1
packaging==25.0
pandas==2.3.1
propcache==0.3.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pynacl==1.5.0
pyopenssl==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
qrcode==8.2
requests==2.32.4
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
spake2==0.9
tenacity==9.1.2
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
twisted==25.5.0
txaio==25.6.1
txtorcon==24.8.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zipstream-ng==1.8.0
zope-interface==7.2
